requirements.txt
Flask
python-dotenv
Flask-Mail
itsdangerous